Vendor typeguard 2.13.3 (#142)
* Vendor typeguard 2.13.3 in its original form.

This package will however need to be edited to work in parallel with other installations of typeguard.

* Inject code theoretically compatible with another injection from another typeguard version.

* Allow typeguard to be installed at same time.

If typeguard is not installed separately, process the packages listed under --typeguard-packages using the vendored typeguard. Otherwise, process only what is listed in --stp-typeguard-packages and leave the --typeguard-packages option to be processed by the installed version of typeguard. To avoid the ambiguity of double processing, an exception is raised if a package is listed in both options.

* Remove namespace changes to vendored code, keep only the updated import change.

This partially reverts commit ea012979c8b9eff99655b3eab8e5cada1e9e4e58.

* Use external typeguard if compatible

* Remove accidentally commited config file

* Add excludes to all static tool configs

* Update github build workflow

* Fix type annotation incompatible with Python 3.8

* Add some more tests to the github build workflow

* Update reference to typechecked decorator in notebook

* Update docs to recommend safer --stp-typeguard-packages pytest option

* add documentation

---------

Co-authored-by: nanne-aben <47976799+nanne-aben@users.noreply.github.com>